requirements/dev.txt
-r base.txt
django-debug-toolbar>=5.0,<6.0
django-extensions>=3.2,<4.0
ipdb>=0.13,<1.0
